fix:(review) outage-class + mechanical fixes from adversarial review
Outage-class: ingester rdkafka gains zstd feature (deployed consumer could not decompress producer output); score-computer treats target_rps=0 as max-rate (MaxRateOnly flag, peak from observed TPS1S) instead of empty wave schedule; validator invariants-mode score uses matched-fill denominator (no uint64 underflow when unmatched>matched); WorkloadSpec.published_at_unix_ns stamp + worker-side WORKLOAD_SPEC_MAX_AGE_S skip (default 300s, 0=not-stale back-compat) defuses stale specs on re-leased partitions.

Mechanical: compression-level filter -1..=12 (librdkafka bounds, also covers ebpf-latency); ViolationCount uint64; violation-example cap per type; minT7 = min over all responses (equivalence oracle updated); REST/WS inflight add/sub pairing + FIX error-path exact-count sub; WS write moves frame.bytes; legacy leaderboard:global ZSET startup DEL; live:latest compare-and-set; skip unchanged live broadcasts; parse-warn rate limit; frontend liveMetrics eviction; active-session query per-session terminal filter; session_band_partition div_ceil + all-partitions property test.

All suites green (one flaky bot-fleet lib test observed passing on rerun — watch item).

+++ services/correctness-validator/internal/store/postgres.go
@@ -216,7 +216,7 @@ SELECT contestant_id, valid_fills, total_fills, correctness_score, violation_cou
 	if err != nil {
 		return ev, false, fmt.Errorf("load score: %w", err)
 	}
-	ev.ValidFills, ev.TotalFills, ev.ViolationCount, ev.ComputedAtNS = uint64(valid), uint64(total), uint32(vcount), uint64(computed)
+	ev.ValidFills, ev.TotalFills, ev.ViolationCount, ev.ComputedAtNS = uint64(valid), uint64(total), uint64(vcount), uint64(computed)
 	ev.SentCount, ev.AckedCount, ev.MatchedCount = uint64(sent), uint64(acked), uint64(matched)
 	return ev, true, nil
 }
